pyproject.toml
[project]
name = "driving-scene-generator"
version = "1.0.0"
description = "Automated pipeline using VLMs to generate structured scene descriptions from autonomous driving images"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pillow>=10.0",
    "tqdm>=4.65",
    "python-dotenv>=1.0",
    "bert-score>=0.3.13",
    "sentence-transformers>=2.2",
    "torch>=2.0",
    "pandas>=2.0",
    "tabulate>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[project.scripts]
dsg = "src.pipeline:main"

[tool.ruff]
line-length = 100
target-version = "py311"
